feat: add migrations
This includes all of the initial set of migrations for pokemons,
generations, types, abilities, users, and access tokens.

The `pokemons.next_evolution_id` is somewhat advanced, in that it is a
recursive reference to another record *on the same table*. My intention
is that different evolutions are stored as entirely distinct pokemon, as
their abilities and stats will likely differ. You can then reference the
next evolution of a given pokemon by their primary key ID. You can find
the highest evolution of a pokemon by looking at pokemon that have
`null` for their `next_evolution_id`.