docs(readme): fix env section in setup instructions, style changes
The change @Gilbert made had a mistake. `cp .env.example .env` is a
command that will copy the first file to the second, so no further steps
required.

Moved the new section about the ApenAI key to directly after the
"environment" section, because that makes sense to me. Also made the
text a little shorter.